tests/wpt/tests/css/css-gaps/parsing/rule-break-computed.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Gaps: *-rule-break getComputedStyle()</title>
  <link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#break">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/computed-testcommon.js"></script>
  <script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  <script>
    const properties = ["column-rule-break", "row-rule-break", "rule-break"];
    for (const property of properties) {
      test_computed_value(property, "none");
      test_computed_value(property, "spanning-item");
      test_computed_value(property, "intersection");
      test_computed_value(property, "auto");
    }
  </script>
</body>
</html>
